Make tests run on both Scala and Dotty.
