+++ core/src/main/scala/better/files/Main.scala
@@ -1,8 +1,7 @@
 package better.files
 
 object Main {
-  def main(args: Array[String]) {
+  def main(args: Array[String]): Unit = {
     file".".listRelativePaths.foreach(println)
   }
-
 }
