+++ core/src/test/scala/better/files/FileSpec.scala
@@ -611,7 +611,7 @@ class FileSpec extends DottyTests {
 }
 
 object Main {
-  def main(args: Array[String]) {
+  def main(args: Array[String]): Unit = {
     new FileSpec().printResults()
   }
 }
